refactor: modularize chart rendering
